openml-generic-python: Fix the models for classification
The model valid_classifier contains bug, and the model second_valid_classifier provides a valid logic.
However, the tests uses the valid_classifier. Therefore, the valid_classifier should be the one that provides a valid logic.

Swap the models' name resolved this problem.
