+++ openml-generic-python/src/test/resources/second_valid_classifier/classifier.py
@@ -1,19 +1,25 @@
 import sys,math
 from ClassifierApi.classifier import ClassifierBase
+import random
+import numpy
 
 class Classifier(ClassifierBase):
 
     num_features=4
     num_classes=3
 
+    def normalize(self, v):
+        norm = numpy.linalg.norm(v, ord=1)
+        if norm == 0:
+            norm = numpy.finfo(v.dtype).eps
+        return v / norm
+
     def classify_instance(self, instance):
-        return self.num_classes-1 - int(sum(instance) % self.num_classes)
+        return int(sum(instance) % self.num_classes)
 
     def getClassDistribution_instance(self, instance):
-        i = self.classify_instance(instance)
-        w = 0.6
-        d = [(1 - w) / (self.num_classes - 1)] * self.num_classes
-        d[i] = w
+        random.seed(instance[0])
+        d = self.normalize(random.sample(range(1, 100), self.num_classes))
         return d
 
     def validate(self, instances):
@@ -22,7 +28,6 @@ class Classifier(ClassifierBase):
                 raise Exception('Instance must be an array!')
             if len(instance) != self.num_features:
                 raise Exception('Incorrect number of features in instance! Got {}. Expected {}.'.format(len(instance), self.num_features))
-
 
     def classify(self, instances):
         self.validate(instances)
