+++ pom.xml
@@ -79,7 +79,7 @@
         <assertj.version>3.7.0</assertj.version>
         <jackson.version>2.6.7</jackson.version>
         <jackson-databind.version>2.6.7</jackson-databind.version>
-        <openml-api.version>1.0.2</openml-api.version>
+        <openml-api.version>1.0.3-bf_pulsedev_28610_openmldatarobotclassify-SNAPSHOT</openml-api.version>
         <jep.version>3.7.0</jep.version>
     </properties>
